refine sync
