+++ bob/playbook/build_internal.go
@@ -7,9 +7,7 @@ import (
 	"sync"
 
 	"github.com/benchkram/bob/bobtask"
-	"github.com/benchkram/bob/bobtask/hash"
 	"github.com/benchkram/bob/pkg/boblog"
-	"github.com/benchkram/bob/pkg/store"
 	"github.com/benchkram/errz"
 )
 
@@ -63,9 +61,8 @@ func (p *Playbook) build(ctx context.Context, task *bobtask.Task) (err error) {
 			hashIn, err := task.HashIn()
 			errz.Fatal(err)
 
-			if p.enableCaching && p.remoteStore != nil {
-				syncFromRemoteToLocal(ctx, p.remoteStore, p.localStore, hashIn)
-			}
+			// download artifact if it exists on the remote
+			p.downloadArtifact(ctx, hashIn)
 
 			success, err := task.ArtifactExtract(hashIn)
 			errz.Fatal(err)
@@ -143,18 +140,4 @@ func (p *Playbook) build(ctx context.Context, task *bobtask.Task) (err error) {
 	boblog.Log.V(1).Info(fmt.Sprintf("%-*s\t%s", p.namePad, coloredName, "..."+state.Short()))
 
 	return nil
-}
-
-// syncFromRemoteToLocal syncs the artifact from the remote store to the local store.
-func syncFromRemoteToLocal(ctx context.Context, remote store.Store, local store.Store, a hash.In) {
-	err := store.Sync(ctx, remote, local, a.String())
-	if errors.Is(err, store.ErrArtifactAlreadyExists) {
-		boblog.Log.V(5).Info(fmt.Sprintf("artifact already exists locally [artifactId: %s]. skipping...", a.String()))
-	} else if errors.Is(err, store.ErrArtifactNotFoundinSrc) {
-		boblog.Log.V(5).Info(fmt.Sprintf("failed to sync from remote to local [artifactId: %s]", a.String()))
-	} else if err != nil {
-		boblog.Log.V(1).Error(err, fmt.Sprintf("failed to sync from remote to local [artifactId: %s]", a.String()))
-	}
-
-	boblog.Log.V(5).Info(fmt.Sprintf("synced from remote to local [artifactId: %s]", a.String()))
 }
